+++ implementation/js/src/reusable/widgets/paginator/index.js
@@ -43,7 +43,7 @@
 			],
 
 			initialize: function(options){
-				this._options = _.extend({
+				this.options = _.extend({
 					pageWindowSize: 5,
 				},options);
 				//if options.target, link to its 'view:page-changed' event
@@ -55,23 +55,23 @@
 				});
 			},
 			onShow: function(){
-				//this.trigger('view:reconfigure', this._options);
+				//this.trigger('view:reconfigure', this.options);
 			},
 			onReconfigure: function(options){
-				_.extend(this._options, options);
+				_.extend(this.options, options);
 				//use options.currentPage, totalPages to build config data - atFirstPage, atLastPage, pages[{number:..., isCurrent:...}]
 				//calculate currentWindow dynamically
-				this._options.currentWindow = Math.ceil(this._options.currentPage/this._options.pageWindowSize);
+				this.options.currentWindow = Math.ceil(this.options.currentPage/this.options.pageWindowSize);
 				var config = {
-					atFirstPage: this._options.currentPage === 1,
-					atLastPage: this._options.currentPage === this._options.totalPages,
-					atFirstWindow: this._options.currentWindow === 1,
-					atLastWindow: this._options.currentWindow === Math.ceil(this._options.totalPages/this._options.pageWindowSize),
-					pages: _.reduce(_.range(1, this._options.totalPages + 1), function(memo, pNum){
-						if(pNum > (this._options.currentWindow - 1) * this._options.pageWindowSize && pNum <= this._options.currentWindow * this._options.pageWindowSize)
+					atFirstPage: this.options.currentPage === 1,
+					atLastPage: this.options.currentPage === this.options.totalPages,
+					atFirstWindow: this.options.currentWindow === 1,
+					atLastWindow: this.options.currentWindow === Math.ceil(this.options.totalPages/this.options.pageWindowSize),
+					pages: _.reduce(_.range(1, this.options.totalPages + 1), function(memo, pNum){
+						if(pNum > (this.options.currentWindow - 1) * this.options.pageWindowSize && pNum <= this.options.currentWindow * this.options.pageWindowSize)
 							memo.push({
 								number: pNum,
-								isCurrent: pNum === this._options.currentPage
+								isCurrent: pNum === this.options.currentPage
 							});
 						return memo;
 					}, [], this)
@@ -82,7 +82,7 @@
 			actions: {
 				goToPage: function($btn, e){
 					var page = $btn.data('page');
-					if(page === this._options.currentPage) return;
+					if(page === this.options.currentPage) return;
 
 					this.trigger('view:change-page', page);
 				},
@@ -90,44 +90,44 @@
 					this.trigger('view:change-page', 1);
 				},
 				goToLastPage: function($btn, e){
-					this.trigger('view:change-page', this._options.totalPages);
+					this.trigger('view:change-page', this.options.totalPages);
 				},
 				//Skipped atm.../////////////////////////
 				// goToAdjacentPage: function($btn, e){
-				// 	var pNum = this._options.currentPage;
+				// 	var pNum = this.options.currentPage;
 				// 	var op = $btn.data('page');
 				// 	if(op === '+')
 				// 		pNum ++;
 				// 	else
 				// 		pNum --;
 
-				// 	if(pNum < 1 || pNum > this._options.totalPages) return;
-				// 	if(pNum > this._options.currentWindow * this._options.pageWindowSize) this._options.currentWindow ++;
-				// 	if(pNum <= (this._options.currentWindow - 1) * this._options.pageWindowSize) this._options.currentWindow --;
+				// 	if(pNum < 1 || pNum > this.options.totalPages) return;
+				// 	if(pNum > this.options.currentWindow * this.options.pageWindowSize) this.options.currentWindow ++;
+				// 	if(pNum <= (this.options.currentWindow - 1) * this.options.pageWindowSize) this.options.currentWindow --;
 				// 	this.trigger('view:change-page', pNum);
 				// },
 				/////////////////////////////////////////
 				goToAdjacentWindow: function($btn, e){
-					var pWin = this._options.currentWindow;
+					var pWin = this.options.currentWindow;
 					var op = $btn.data('window');
 					if(op === '+')
 						pWin ++;
 					else
 						pWin --;
 
-					if (pWin < 1 || pWin > Math.ceil(this._options.totalPages/this._options.pageWindowSize)) return;
-					this.trigger('view:change-page', (pWin == 1) ? 1 : (pWin-1) * this._options.pageWindowSize + 1);
+					if (pWin < 1 || pWin > Math.ceil(this.options.totalPages/this.options.pageWindowSize)) return;
+					this.trigger('view:change-page', (pWin == 1) ? 1 : (pWin-1) * this.options.pageWindowSize + 1);
 				}
 			},
 			//////Can be overriden in options to add extra params///////
 			onChangePage: function(pNum){
 				//use the overriden version (see the stub impl below for what to override)
-				if(this._options.onChangePage)
-					return this._options.onChangePage.call(this, pNum);
+				if(this.options.onChangePage)
+					return this.options.onChangePage.call(this, pNum);
 
 				//use just a default stub implementation
-				if(this._options.target) 
-					this._options.target.trigger('view:load-page', {
+				if(this.options.target) 
+					this.options.target.trigger('view:load-page', {
 						page: pNum
 						//add more params/querys
 					});